+++ src/locales/nb_NO/location/state.ts
@@ -13,6 +13,6 @@ export default [
   'Vestland',
   'Telemark',
   'Troms',
-  'Trööndelage',
+  'Trøndelag',
   'Østfold',
 ];