+++ test/vitest-extensions.ts
@@ -17,10 +17,33 @@ expect.extend({
           : `No duplicate values in [${received.join(', ')}]`,
     };
   },
+  toStartWith(actual: unknown, prefix: string) {
+    const { isNot } = this;
+
+    return {
+      pass: typeof actual === 'string' && actual.startsWith(prefix),
+      actual,
+      expected: `${prefix}...`,
+      message: () =>
+        isNot
+          ? `String did start with the "${prefix}".`
+          : `String did not start with the "${prefix}".`,
+    };
+  },
 });
 
 interface CustomMatchers {
+  /**
+   * Expects that a list of elements does not contain any duplicate entries.
+   */
   toContainDuplicates(): void;
+
+  /**
+   * Expects that a string has the provided prefix.
+   *
+   * @param prefix The prefix to check for.
+   */
+  toStartWith(prefix: string): void;
 }
 
 declare module 'vitest' {